On error, fail the send Future
Before using Futures, errors were handled using the `error` event
provided by Role::EventEmitter. But this is not very IO::Async-ish,
since we already have users `get`ting Futures, and if we simply
fire an error event, that Future will never be complete.

This patch changes that so that the Future created by the `send`
method is `done` on success and `fail`ed on error, and the user is
free to provide a `catch` block for it.

This means `get`ting a Future returned by `send` could now possibly
throw an exception.

This patch also registers a dummy listener to the `error` event, so
that Role::EventEmitter doesn't die when the event is triggered. The
event is therefore now turned into something the user can register to
for additional information, instead of something the user _must_
register to for safe execution.